requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
pillow==12.0.0
python-multipart==0.0.20
svgwrite==1.4.3
pydantic==2.12.5
